AWS/requirements.txt
aws-cdk-lib==2.43.1